Extract TrustedRouter model catalog defaults
