+++ docs/CODE_QUALITY_AUDIT.md
@@ -10,7 +10,7 @@ The architecture is moving in the right direction: core state is value typed, pe
 
 | Component | Grade | Notes |
 | --- | --- | --- |
-| `QuillCodeCore` | A- | Stable models, canonical TrustedRouter IDs, branded display names, and compatibility decoding. Keep pushing presentation-only naming into helpers instead of scattering strings. |
+| `QuillCodeCore` | A | Stable value models, focused TrustedRouter default/catalog ownership, canonical IDs, branded display names, and compatibility decoding. Keep future provider-specific behavior out of the general domain model file. |
 | `QuillCodeAgent` | A- | Runtime/tool loop is well covered and keeps tool feedback hidden from user transcript surfaces. Next grade step is richer retry/cancellation telemetry. |
 | `QuillCodeTools` | A- | Shell/file/git/MCP executors are bounded and testable. Git and MCP files are necessarily dense; keep extracting parsers/policies when behavior grows. |
 | `QuillCodeSafety` | A- | Small, explicit policy layer. Needs more production prompt telemetry once live Auto reviewer tuning begins. |
@@ -30,7 +30,9 @@ The architecture is moving in the right direction: core state is value typed, pe
 | `Sources/quill-code-desktop/QuillCodeDesktopApp.swift` | A- | App scene composition is now small and declarative. Keep it limited to window/menu-bar wiring and root-view routing. |
 | `Sources/quill-code-desktop/QuillCodeDesktopController.swift` | A- | Desktop controller is now mostly UI/workspace routing. Pasteboard feedback and project-import resolution now live in focused coordinators; keep future desktop protocol/workflow details out of the controller. |
 | `Sources/QuillCodeAgent/Agent.swift` | A- | Good test coverage; keep tool continuation limits and transcript filtering explicit. |
-| `Sources/QuillCodeCore/Models.swift` | A- | Central source of truth for model IDs, branding, and compatibility. Watch for model/persistence surface bloat. |
+| `Sources/QuillCodeCore/Models.swift` | A | General domain models only; TrustedRouter defaults/catalog records now live in focused core files. Watch for persistence or provider-specific behavior trying to drift back in. |
+| `Sources/QuillCodeCore/TrustedRouterDefaults.swift` | A | Central source of truth for TrustedRouter IDs, aliases, branded model names, fallback catalog rows, and catalog normalization. |
+| `Sources/QuillCodeCore/ModelInfo.swift` | A | Small catalog value records and sort-key semantics with no app/runtime dependency. |
 
 ## Changes From This Pass
 
@@ -40,6 +42,19 @@ The architecture is moving in the right direction: core state is value typed, pe
 - Refactored model-category construction to compute favorite IDs once and pass a `Set` through option building instead of recomputing favorites for every model.
 - Updated the Playwright harness to preserve branded labels after model selection.
 - Fixed stale decisions documentation that still described recurring automation as deferred.
+
+## 2026-06-23 Core Model Catalog Ownership Pass
+
+Overall grade after this slice: **A core cohesion, A model-default ownership**.
+
+`Models.swift` was still carrying provider-specific TrustedRouter defaults, catalog records, and sort policy beside general thread/config/domain models. The behavior was correct, but the file mixed broad app domain state with one provider's catalog policy, making future provider/model-picker work easier to scatter.
+
+Code quality changes:
+
+- Moved `ModelInfo` and `ModelSortKey` into `ModelInfo.swift`.
+- Moved `TrustedRouterDefaults` into `TrustedRouterDefaults.swift`, keeping Nike 1.0, Prometheus 1.0, aliases, fallback catalog rows, and catalog normalization together.
+- Reduced `Models.swift` by keeping it focused on general domain models and app config/auth records.
+- Added a parity gate that prevents model catalog records, sort keys, TrustedRouter defaults, and model branding copy from drifting back into `Models.swift`.
 
 ## Current Refactor Priority
