+++ docs/DECISIONS.md
@@ -6,6 +6,7 @@
 - TrustedRouter prompt/message construction lives in `TrustedRouterPromptBuilder`, not the network client. The builder owns system prompt copy, tool schema formatting, project instruction context, memory context, message-history projection, tool-output projection, and the history limit. `TrustedRouterLLMClient` owns SDK calls and stream collection.
 - TrustedRouter API-key resolution lives in `TrustedRouterAPIKeyResolver`. Developer override precedence, whitespace trimming, session-store fallback, and the actionable missing-key error should stay there so action and safety clients cannot drift.
 - TrustedRouter action and safety transports are separate files. `TrustedRouterLLMClient` owns action streaming; `TrustedRouterSafetyModelClient` owns Auto-review model calls. They share API-key resolution through `TrustedRouterAPIKeyResolver` and JSON response payloads through `TrustedRouterChatParameters` so neither transport depends on the other.
+- TrustedRouter model defaults and catalog normalization live in focused core files, not the general `Models.swift` domain bucket. `TrustedRouterDefaults.swift` owns provider IDs, aliases, bundled fallback rows, Nike 1.0/Prometheus 1.0 branding, and sort/category policy; `ModelInfo.swift` owns the catalog value records.
 - Tool-router feature families should delegate before they sprawl. `ToolRouter` owns the common shell/file/patch entry point and composes git definitions from `GitToolCallDispatcher`; local git, GitHub PR, and worktree tool-call argument mapping lives in that dispatcher. This keeps the shared router small while preserving one public `ToolRouter.execute` API for the agent runtime.
 - Shell request policy is a tool-family concern. `ShellToolCallDispatcher` owns `host.shell.run` definitions, cwd containment, timeout bounds, environment override validation, and `ShellExecutionRequest` construction; `ToolRouter` only delegates shell calls before handling file and patch primitives.
